long замеен на BigInteger, BigInteger не дружит с hibernate

+++ src/main/java/com/practice/taskServer/data/dto/AcceptTaskDTO.java
@@ -2,13 +2,15 @@ package com.practice.taskServer.data.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.math.BigInteger;
+
 //объект для представления отчета с результатом выполнения задачи
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AcceptTaskDTO {
 
     public String name;
     public String worker;
-    public long number;
+    public BigInteger number;
     public boolean isPrime;
 
     public AcceptTaskDTO() {;
