+++ src/main/java/com/practice/taskServer/data/dto/GetTaskDTO.java
@@ -2,16 +2,18 @@ package com.practice.taskServer.data.dto;
 
 import com.practice.taskServer.data.entity.TaskEntity;
 
+import java.math.BigInteger;
+
 //объект для представления информации о задаче
 public class GetTaskDTO {
 
     public String name;
-    public long number;
+    public BigInteger number;
 
     public GetTaskDTO() {
     }
 
-    public GetTaskDTO(String name, long number) {
+    public GetTaskDTO(String name, BigInteger number) {
         this.name = name;
         this.number = number;
     }
